requirements/default.txt
Django>=1.6
djangorestframework>=2.4.3
django-oauth-toolkit==0.10.0
django-otp==0.3.5
httplib2==0.9.2
oauthlib==1.0.3
pytz==2016.4
six==1.10.0
twilio==5.4.0